Use adm-zip and move process back to desktop_scripts for simplicity
